+++ braintree/payment_instrument_type.py
@@ -9,3 +9,6 @@ class PaymentInstrumentType():
     AmexExpressCheckoutCard = "amex_express_checkout_card"
     VenmoAccount = "venmo_account"
     UsBankAccount = "us_bank_account"
+    IdealPayment = "ideal_payment"
+    VisaCheckoutCard = "visa_checkout_card"
+    MasterpassCard = "masterpass_card"