+++ braintree/payment_method_gateway.py
@@ -9,6 +9,8 @@ from braintree.android_pay_card import AndroidPayCard
 from braintree.amex_express_checkout_card import AmexExpressCheckoutCard
 from braintree.venmo_account import VenmoAccount
 from braintree.us_bank_account import UsBankAccount
+from braintree.visa_checkout_card import VisaCheckoutCard
+from braintree.masterpass_card import MasterpassCard
 from braintree.unknown_payment_method import UnknownPaymentMethod
 from braintree.error_result import ErrorResult
 from braintree.exceptions.not_found_error import NotFoundError
@@ -148,6 +150,10 @@ class PaymentMethodGateway(object):
             return VenmoAccount(self.gateway, response["venmo_account"])
         elif "us_bank_account" in response:
             return UsBankAccount(self.gateway, response["us_bank_account"])
+        elif "visa_checkout_card" in response:
+            return VisaCheckoutCard(self.gateway, response["visa_checkout_card"])
+        elif "masterpass_card" in response:
+            return MasterpassCard(self.gateway, response["masterpass_card"])
         elif "payment_method_nonce" in response:
             return PaymentMethodNonce(self.gateway, response["payment_method_nonce"])
         elif "success" in response: