+++ braintree/transaction.py
@@ -28,6 +28,9 @@ from braintree.descriptor import Descriptor
 from braintree.risk_data import RiskData
 from braintree.three_d_secure_info import ThreeDSecureInfo
 from braintree.us_bank_account import UsBankAccount
+from braintree.ideal_payment import IdealPayment
+from braintree.visa_checkout_card import VisaCheckoutCard
+from braintree.masterpass_card import MasterpassCard
 from braintree.facilitator_details import FacilitatorDetails
 from braintree.payment_instrument_type import PaymentInstrumentType
 
@@ -537,6 +540,8 @@ class Transaction(Resource):
                     "venmo_sdk_session",
                     "payee_email",
                     "skip_advanced_fraud_checking",
+                    "skip_avs",
+                    "skip_cvv",
                     {
                         "paypal": [
                             "payee_email",
@@ -630,6 +635,8 @@ class Transaction(Resource):
             self.europe_bank_account_details = EuropeBankAccount(gateway, attributes.pop("europe_bank_account"))
         if "us_bank_account" in attributes:
             self.us_bank_account = UsBankAccount(gateway, attributes.pop("us_bank_account"))
+        if "ideal_payment" in attributes:
+            self.ideal_payment_details = IdealPayment(gateway, attributes.pop("ideal_payment"))
         if "apple_pay" in attributes:
             self.apple_pay_details = ApplePayCard(gateway, attributes.pop("apple_pay"))
         if "coinbase_account" in attributes:
@@ -640,6 +647,10 @@ class Transaction(Resource):
             self.amex_express_checkout_card_details = AmexExpressCheckoutCard(gateway, attributes.pop("amex_express_checkout_card"))
         if "venmo_account" in attributes:
             self.venmo_account_details = VenmoAccount(gateway, attributes.pop("venmo_account"))
+        if "visa_checkout_card" in attributes:
+            self.visa_checkout_card_details = VisaCheckoutCard(gateway, attributes.pop("visa_checkout_card"))
+        if "masterpass_card" in attributes:
+            self.masterpass_card_details = MasterpassCard(gateway, attributes.pop("masterpass_card"))
         if "customer" in attributes:
             self.customer_details = Customer(gateway, attributes.pop("customer"))
         if "shipping" in attributes: