+++ tests/integration/test_transaction.py
@@ -899,6 +899,39 @@ class TestTransaction(unittest.TestCase):
         self.assertIsInstance(transaction.risk_data, RiskData)
         self.assertEqual(transaction.risk_data.id, None)
 
+    def test_sale_with_skip_cvv_option_set(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": CreditCardNumbers.Visa,
+                "expiration_date": "05/2009"
+            },
+            "options": {
+                "skip_cvv": True
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEqual(transaction.cvv_response_code, "B")
+
+    def test_sale_with_skip_avs_option_set(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": CreditCardNumbers.Visa,
+                "expiration_date": "05/2009"
+            },
+            "options": {
+                "skip_avs": True
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEqual(transaction.avs_error_response_code, None)
+        self.assertEqual(transaction.avs_street_address_response_code, "B")
+
     def test_validation_error_on_invalid_custom_fields(self):
         result = Transaction.sale({
             "amount": TransactionAmounts.Authorize,
@@ -3487,3 +3520,46 @@ class TestTransaction(unittest.TestCase):
         self.assertEqual(result.transaction.shipping_details.first_name, address.first_name)
         self.assertEqual(result.transaction.billing_details.first_name, address.first_name)
         self.assertEqual(result.transaction.customer_details.first_name, customer.first_name)
+
+    def test_sale_transacts_ideal_payment(self):
+        valid_id = TestHelper.generate_valid_ideal_payment_id()
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "order_id": "ABC123",
+            "merchant_account_id": "ideal_merchant_account",
+            "payment_method_nonce": valid_id,
+            "options": {
+                "submit_for_settlement": True,
+            },
+        })
+
+        self.assertTrue(result.is_success)
+
+        self.assertRegexpMatches(result.transaction.id, r'^\w{6,}$')
+        self.assertEqual(result.transaction.type, "sale")
+        self.assertEqual(result.transaction.payment_instrument_type, PaymentInstrumentType.IdealPayment)
+        self.assertEqual(result.transaction.amount, Decimal(TransactionAmounts.Authorize))
+        self.assertEqual(result.transaction.status, Transaction.Status.Settled)
+        self.assertRegexpMatches(result.transaction.ideal_payment_details.ideal_payment_id, r"^idealpayment_\w{6,}")
+        self.assertRegexpMatches(result.transaction.ideal_payment_details.ideal_transaction_id, r"^\d{16,}$")
+        self.assertEqual(result.transaction.ideal_payment_details.image_url[:8], 'https://')
+        self.assertNotEqual(result.transaction.ideal_payment_details.masked_iban, None)
+        self.assertNotEqual(result.transaction.ideal_payment_details.bic, None)
+
+    def test_failed_sale_non_complete_ideal_payment(self):
+        non_complete_id = TestHelper.generate_valid_ideal_payment_id("3.00")
+        result = Transaction.sale({
+            "amount": "3.00",
+            "order_id": "ABC123",
+            "merchant_account_id": "ideal_merchant_account",
+            "payment_method_nonce": non_complete_id,
+            "options": {
+                "submit_for_settlement": True,
+            },
+        })
+        error_codes = [
+            error.code for error in result.errors.for_object("transaction").on("payment_method_nonce")
+        ]
+
+        self.assertFalse(result.is_success)
+        self.assertTrue(ErrorCodes.Transaction.IdealPaymentNotComplete in error_codes)