+++ tests/unit/test_credit_card.py
@@ -79,7 +79,8 @@ class TestCreditCard(unittest.TestCase):
         CreditCard.find(None)
 
     def test_multiple_verifications_sort(self):
-        verification1 = {"created_at": datetime.datetime(2014, 11, 18, 23, 20, 20), "id":123}
-        verification2 = {"created_at": datetime.datetime(2014, 11, 18, 23, 20, 21), "id":456}
+        verification1 = {"created_at": datetime.datetime(2014, 11, 18, 23, 20, 20), "id": 123, "amount": "0.00"}
+        verification2 = {"created_at": datetime.datetime(2014, 11, 18, 23, 20, 21), "id": 456, "amount": "1.00"}
         credit_card = CreditCard(Configuration.gateway(), {"verifications": [verification1, verification2]})
         self.assertEqual(456, credit_card.verification.id)
+        self.assertEqual(1.00, credit_card.verification.amount)